# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  internal_server_endpoints.Book:
    properties:
      author:
        type: string
      created_at:
        type: string
      id:
        type: string
      metadata_complete:
        type: boolean
      page_count:
        type: integer
      status:
        type: string
      status_reason:
        type: string
      structure_complete:
        type: boolean
      structure_failed:
        type: boolean
      title:
        type: string
    type: object
...
  internal_server_endpoints.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_server_endpoints.IngestRequest:
    properties:
      author:
        type: string
      pdf_paths:
        items:
          type: string
        type: array
      stitch:
        type: boolean
      title:
        type: string
    type: object
  internal_server_endpoints.IngestResponse:
    properties:
      author:
        type: string
      book_id:
        type: string
      job_id:
        type: string
      process_job_id:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
  internal_server_endpoints.ListBooksResponse:
    properties:
      books:
        items:
          $ref: '#/definitions/internal_server_endpoints.Book'
        type: array
    type: object
...
  internal_server_endpoints.RerunTocResponse:
    properties:
      job_id:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/jackzampolin/shelf
  description: Book digitization pipeline API for managing books, jobs, and processing.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Shelf API
  version: "1.0"
paths:
...
  /api/books:
    get:
      description: List all books in the library
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_server_endpoints.ListBooksResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
      summary: List books
      tags:
      - books
...
  /api/books/{book_id}/rerun-toc:
    post:
      description: Reset ToC state and rerun the ToC finder and extractor agents
      parameters:
      - description: Book ID
        in: path
        name: book_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/internal_server_endpoints.RerunTocResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "412":
          description: Not enough pages have OCR complete
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
      summary: Rerun ToC finder and extractor
      tags:
      - books
...
  /api/books/ingest:
    post:
      consumes:
      - application/json
      description: Ingest PDF files as a new book and start processing
      parameters:
      - description: Ingest request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_server_endpoints.IngestRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/internal_server_endpoints.IngestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_server_endpoints.ErrorResponse'
      summary: Ingest book scans
      tags:
      - books
...
swagger: "2.0"